Move definition of Half to out of native backend
